% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{ashsmooth.gaus}
\alias{ashsmooth.gaus}
\title{Estimate the underlying mean function from noisy Gaussian data}
\usage{
ashsmooth.gaus(x, sigma = NULL, v.est = FALSE, joint = FALSE,
  v.basis = FALSE, post.var = FALSE, filter.number = 1,
  family = "DaubExPhase", jash = FALSE, SGD = TRUE, weight = 0.5,
  min.var = 1e-08, ashparam = list())
}
\arguments{
\item{filter.number,}{family: wavelet basis to be used, as in \code{wavethresh}}

\item{sigma:}{a vector of standard deviations. Can be provided if known or estimated beforehand.}

\item{v.est:}{bool, indicating if variance estimation should be performed instead.}

\item{joint:}{bool, indicating if results of mean and variance estimation should be returned together.}

\item{v.basis:}{bool, indicating if the same wavelet basis should be used for variance estimation as mean estimation. If false, defaults to Haar basis for variance estimation (this is much faster than other bases).}

\item{post.va:}{bool, indicating if the posterior variance should be returned for the mean and/or variance estiamtes.}

\item{jash:}{indicates if the prior from method JASH should be used. This will typically provide better variance estimates (especially for nonsmooth variance functions), at the cost of computational efficiency. Defaults to FALSE.}

\item{SGD:}{bool, indicating if stochastic gradient descent should be used in the EM. Only applicable if jash=TRUE.}

\item{weight:}{optional parameter used in estimating overall variance. Only works for Haar basis. Defaults to 0.5. Setting this to 1 might improve variance estimation slightly}

\item{min.var:}{The minimum positive value to be set if the variance estimates are negative.}

\item{ashparam:}{a list of parameters to be passed to \code{ash}; default values are set by function \code{\link{setAshParam}}.}
}
\value{
\code{ashsmooth.gaus} returns the mean estimate by default, or the variance estimate if \code{v.est} is TRUE. However, if \code{joint} is TRUE, then a list with the following is returned:
\item{mu.res}{a list with the mean estimate and its posterior variance if \code{post.var} is TRUE, or a vector of mean estimates otherwise}
\item{var.res}{a list with the variance estimate and its posterior variance if \code{v.est} is TRUE and \code{post.var} is TRUE, or a vector of variance estimates if only \code{v.est} is TRUE}
}
\description{
This function takes a data vector of length a power of 2 and performs signal denoising using wavelet decomposition and an adaptive shrinkage prior on the wavelet parameters. The data are assumed to be (mostly) independent and Gaussian, but not necessarily identically distributed.
}

