% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{ashsmooth.pois}
\alias{ashsmooth.pois}
\title{Main smoothing procedure for Poisson data. Takes a univariate inhomogeneous Poisson process and estimates its mean intensity}
\usage{
ashsmooth.pois(x, post.var = FALSE, reflect = FALSE, lev = 0,
  log = FALSE, pseudocounts = 0.5, all = FALSE, lm.approx = FALSE,
  cxx = TRUE, ashparam = list())
}
\arguments{
\item{x:}{a vector of Poisson counts of a length a power of 2}

\item{post.var:}{bool, indicates if the posterior variance should be returned}

\item{reflect:}{bool, indicates if the signals should be reflected; otherwise periodicity is assumed}

\item{lev:}{integer from 0 to J-1, indicating primary level of resolution. Should NOT be used (ie shrinkage is performed at all resolutions) unless there is good reason to do so.}

\item{log:}{bool, determines if smoothed signal is returned on log scale or not}

\item{pseudocounts:}{bool, a number to be added to counts. Passed to \code{glm.approx}}

\item{all:}{bool, indicates if pseudocounts should be added too all entries or only cases when either number of successes or number of failures (but not both) is 0. Passed to \code{glm.approx}}

\item{lm.approx:}{bool, indicates if a WLS shuold be fitted instead of GLM. Passed to \code{glm.approx}}

\item{cxx:}{bool, indicates if C++ code should be used to create TI tables.}

\item{ashparam:}{a list of parameters to be passed to \code{ash}; default values are set by function \code{\link{setAshParam}}.}
}
\value{
\code{ashsmooth.pois} returns the mean estimate by default, or the variance estimate if \code{post.var} is TRUE
}
\description{
Main smoothing procedure for Poisson data. Takes a univariate inhomogeneous Poisson process and estimates its mean intensity
}
\examples{
n=2^10
t=1:n/n
spike.f=function(x) (0.75*exp(-500*(x-0.23)^2)+1.5*exp(-2000*(x-0.33)^2)+3*exp(-8000*(x-0.47)^2)+2.25*exp(-16000*(x-0.69)^2)+0.5*exp(-32000*(x-0.83)^2))
mu.s=spike.f(t)
mu.t=0.01+mu.s
X.s=rpois(n,mu.t)
mu.est=ashsmooth.pois(X.s)
plot(mu.t,type="l")
lines(mu.est,col=2)
}

