% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{compute.approx.z}
\alias{compute.approx.z}
\title{compute.approx.z
Compute a vector of logit(p) given a vector of successes and failures, as well as its variance estimates (MLE with approximation at endpoints for mean; a mix of Berkso's estimator and Tukey's estimator for variance)}
\usage{
compute.approx.z(x.s, x.f, bound, eps, pseudocounts, all, indexn = NULL,
  return.p = FALSE)
}
\value{
a list with elements "mu", "var" and optionally "p"
}
\description{
compute.approx.z
Compute a vector of logit(p) given a vector of successes and failures, as well as its variance estimates (MLE with approximation at endpoints for mean; a mix of Berkso's estimator and Tukey's estimator for variance)
}

