% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{compute.res}
\alias{compute.res}
\title{Compute posterior mean and var for log(p), log(q), log(p0/p1) and log(q0/q1)
This function returns posterior means and variances of log(p), log(q), log(p0/p1) and log(q0/q1) as lp, lq, lpratio and lqratio, respectively, where p
is the probability of going left and q=1-p.}
\usage{
compute.res(alpha, log)
}
\arguments{
\item{log:}{indicates if mean estimation is to be performed on the log scale}
}
\value{
a list with elements "lp.mean", "lp.var", "lq.mean", "lq.var"
}
\description{
Compute posterior mean and var for log(p), log(q), log(p0/p1) and log(q0/q1)
This function returns posterior means and variances of log(p), log(q), log(p0/p1) and log(q0/q1) as lp, lq, lpratio and lqratio, respectively, where p
is the probability of going left and q=1-p.
}

