% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMASH.R
\docType{data}
\name{eS}
\alias{eS}
\title{A collection of pre-defined subclone configurations.}
\format{
An object of class \code{list} of length 5.
}
\usage{
eS
}
\description{
A R list containing subclone configurations in matrix form for 1 to 5 subclones. For each matrix, each column corresponds to a subclone and each row corresponds to a variant's allocation across all subclones. For example, the first row of each matrix is a vector of 1's to represent clonal variants, variants present in all subclones.
}
\keyword{datasets}
