% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{getlist.res}
\alias{getlist.res}
\title{For each resolution, performs shrinkage and returns the posterior means and variances in matrix form}
\usage{
getlist.res(res, j, n, zdat, log, shrink, prior, pointmass, nullcheck, gridmult,
  mixsd, VB, mixcompdist)
}
\description{
For each resolution, performs shrinkage and returns the posterior means and variances in matrix form
}

