% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{interleave}
\alias{interleave}
\title{Interleave two vectors}
\usage{
interleave(x, y)
}
\arguments{
\item{x,y:}{two vectors of the same length}
}
\value{
a vector of length twice that of x (or y)
}
\description{
Interleave two vectors
}

