% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{ndwt.mat}
\alias{ndwt.mat}
\title{Computes the non-decimated wavelet transform matrix for a given basis}
\usage{
ndwt.mat(n, filter.number, family)
}
\arguments{
\item{filter.number,family:}{specifies the type of wavelet basis used}

\item{n:}{the sample size. Must be a power of 2}
}
\value{
the NDWT matrix for the specified basis, with the entries squared
}
\description{
Computes the non-decimated wavelet transform matrix for a given basis
}

