% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{ParentTItable}
\alias{ParentTItable}
\title{Produces two TI tables for Poisson counts. One is the standard TI table, and the other are the parent values used to make the TI table}
\usage{
ParentTItable(sig)
}
\arguments{
\item{sig:}{a signal of length a power of 2}
}
\value{
a list of two TI tables in the form of matrices
}
\description{
Produces two TI tables for Poisson counts. One is the standard TI table, and the other are the parent values used to make the TI table
}

