% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{recons.mv}
\alias{recons.mv}
\title{Reconstructs the signal in data space}
\usage{
recons.mv(ls, res, log, n, J)
}
\arguments{
\item{ls:}{estimated total intensity}

\item{res:}{the matrix of wavelet proportions/probabilities}

\item{log:}{bool, indicating if signal reconstruction should be in log space}

\item{n:}{length of data}

\item{J:}{=log2(n)

return reconstructed signal in original data space}
}
\description{
Reconstructs the signal in data space
}

