% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{reflect}
\alias{reflect}
\title{reflects a vector if it has length a power of 2; otherwise extends the vector to have length a power of 2 and then reflects it}
\usage{
reflect(x)
}
\arguments{
\item{x}{an n-vector}
}
\value{
an n-vector containing the indices of the original signal x
}
\description{
reflects a vector if it has length a power of 2; otherwise extends the vector to have length a power of 2 and then reflects it
}

