% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{reverse.gvwave}
\alias{reverse.gvwave}
\title{Reverse wavelet transform a set of posterior variances for wavelet coefficients in TItable format, for Gaussian data.}
\usage{
reverse.gvwave(est, lp, lq = NULL)
}
\arguments{
\item{lp:}{a J by n matrix of estimated variances.}

\item{lq:}{a J by n matrix of complementary variances (=lp).}

\item{est:}{an n-vector. Usually 0.}
}
\value{
reconstructed posterior variance in the original data space.
}
\description{
Reverse wavelet transform a set of posterior variances for wavelet coefficients in TItable format, for Gaussian data.
}

