% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{reverse.pwave}
\alias{reverse.pwave}
\title{Reverse wavelet transform a set of probabilities in TItable format for Poisson data.}
\usage{
reverse.pwave(est, lp, lq = NULL)
}
\arguments{
\item{lp:}{a J by n matrix of probabilities.}

\item{lq:}{a J by n matrix of complementary probabilities.}

\item{est:}{an n-vector. Usually a constant vector with each element equal to the estimated log(total rate).}
}
\value{
reconstructed signal in the original data space.
}
\description{
Reverse wavelet transform a set of probabilities in TItable format for Poisson data.
}

