% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{rshift}
\alias{rshift}
\title{Shift a vector one unit to the right}
\usage{
rshift(x)
}
\arguments{
\item{x:}{a vector}
}
\value{
a vector of the same length as that of x
}
\description{
Shift a vector one unit to the right
}

