% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{setAshParam.gaus}
\alias{setAshParam.gaus}
\title{Set default \code{ash} parameters.}
\usage{
setAshParam.gaus(ashparam)
}
\arguments{
\item{ashparam:}{a list of parameters to be passed to ash.}
}
\description{
Set default \code{ash} parameters.
}
\keyword{internal}

