% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{simplify}
\alias{simplify}
\title{Turns a TItable into a matrix that one can apply glm to (for Poisson denoising).}
\usage{
simplify(dmat)
}
\arguments{
\item{dmat:}{a k by n matrix
return a 2 by nk/2 matrix}
}
\description{
Turns a TItable into a matrix that one can apply glm to (for Poisson denoising).
}

