% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{smash}
\alias{smash}
\alias{smash-package}
\title{Smash: SMoothing using Adaptive SHrinkage}
\description{
This function performs nonparametric regression on univariate Poisson or Gaussian signals
}

