% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{threshold.haar}
\alias{threshold.haar}
\title{This is a modified threshold.wd function from package "wavethresh", designed to perform thresholding for heteroskedastic Gaussian errors when using the Haar basis}
\usage{
\method{threshold}{haar}(vdtable, vtable, levels, type = "hard",
  policy = "universal")
}
\description{
This is a modified threshold.wd function from package "wavethresh", designed to perform thresholding for heteroskedastic Gaussian errors when using the Haar basis
}

