% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{threshold.var}
\alias{threshold.var}
\title{This is a modified threshold.wd function from package "wavethresh", designed to perform thresholding for heteroskedastic Gaussian errors when using non-Haar basis}
\usage{
\method{threshold}{var}(x.w, x.w.v, levels, type = "hard")
}
\description{
This is a modified threshold.wd function from package "wavethresh", designed to perform thresholding for heteroskedastic Gaussian errors when using non-Haar basis
}

