% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{tirtable}
\alias{tirtable}
\title{Produces a TI table containing the log difference between adjacent pairs of data in the same resolution.}
\usage{
tirtable(sig)
}
\arguments{
\item{sig:}{a signal of length a power of 2}
}
\value{
a TI tables in the form of a matrix
}
\description{
Produces a TI table containing the log difference between adjacent pairs of data in the same resolution.
}

