% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{var.smooth}
\alias{var.smooth}
\title{var.smooth}
\usage{
var.smooth(data, data.var, x.var.ini, basis, v.basis, Wl, filter.number, family,
  post.var, prior, pointmass, nullcheck, VB, mixsd, mixcompdist, gridmult, jash,
  weight, J, n, SGD)
}
\value{
"var.est" if posterior variances are not computed, and a list with elements "var.est" and "var.est.var" otherwise
}
\description{
var.smooth
}
\keyword{internal}

