% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMASH.R
\name{vis_GRID}
\alias{vis_GRID}
\title{vis_GRID}
\usage{
vis_GRID(GRID)
}
\arguments{
\item{GRID}{The \code{GRID} object output from \code{grid_ITH_optim}.}
}
\value{
A ggplot object for data visualization
}
\description{
A simple visualization of SMASH's grid of solutions
}
