% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/glm_approx.R
\name{vs}
\alias{vs}
\title{Compute approximately unbiased variance estimates for the estimators for logit(p) when n is small}
\usage{
vs(n, s, f)
}
\arguments{
\item{n:}{number of trials}

\item{s:}{number of successes}

\item{f:}{number of failures}
}
\description{
Compute approximately unbiased variance estimates for the estimators for logit(p) when n is small
}

