% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{wd.D}
\alias{wd.D}
\title{Modified wd() function from package "wavethresh" to only return the detail coefficients}
\usage{
wd.D(data, filter.number = 10, family = "DaubLeAsymm", type = "wavelet",
  bc = "periodic", verbose = FALSE, min.scale = 0, precond = TRUE)
}
\arguments{
\item{data,filter.number,family,type,bc,berbose,min.scale,precond:}{as in wd()}
}
\value{
the detail coefficients of a standard wavelet decomposition
}
\description{
Modified wd() function from package "wavethresh" to only return the detail coefficients
}

