% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{ashsmooth.pois}
\alias{ashsmooth.pois}
\title{Main smoothing procedure for Poisson data. Takes a univariate inhomogeneous Poisson process and estimates its mean intensity}
\usage{
ashsmooth.pois(x, post.var = FALSE, reflect = FALSE, lev = 0,
  log = FALSE, pseudocounts = 0.5, all = FALSE, ashparam = list())
}
\arguments{
\item{x:}{a vector of Poisson counts of a length a power of 2}

\item{post.var:}{bool, indicates if the posterior variance should be returned}

\item{reflect:}{bool, indicates if the signals should be reflected; otherwise periodicity is assumed}

\item{lev:}{integer from 0 to J-1, indicating primary level of resolution. Should NOT be used (ie shrinkage is performed at all resolutions) unless there is good reason to do so.}

\item{log:}{bool, determines if smoothed signal is returned on log scale or not}

\item{pseudocounts:}{bool, a number to be added to counts. Passed to \code{glm.approx}}

\item{all:}{bool, indicates if pseudocounts should be added too all entries or only cases when either number of successes or number of failures (but not both) is 0. Passed to \code{glm.approx}}

\item{ashparam:}{a list of parameters to be passed to \code{ash}; default values are set by function \code{\link{setAshParam}}.}
}
\value{
\code{ashsmooth.pois} returns the mean estimate by default, or the variance estimate if \code{post.var} is TRUE
}
\description{
Main smoothing procedure for Poisson data. Takes a univariate inhomogeneous Poisson process and estimates its mean intensity
}

