% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/documentation.R
\docType{package}
\name{smash}
\alias{smash}
\alias{smash-package}
\title{Smash: Smoothing using Adaptive SHrinkage}
\description{
This package performs nonparametric regression on univariate Poisson or Gaussian data using wavelets. For the Poisson case, the data are assumed to be i.i.d.
from an underlying inhomogeneous mean function that is "smooth". Similarly for the Gaussian case, the data are assumed to be independent with an underlying smooth mean function.
In the Gaussian case, the variances are allowed vary, but are similarly "spatially structured" as with the mean function.
}
\details{
The functions \code{ashsmooth.gaus} and \code{ashsmooth.pois} perform smoothing for Gaussian and Poisson data respectively. The only required input is a vector of length 2^J for some integer J.
Other options include the possibility of returning the posterior variances, specifying a wavelet basis (default is Haar, which performs well in general due to the fact that we used the translation-invariant version)
}
\examples{
n=2^9
t=1:n/n
spike.f=function(x) (0.75*exp(-500*(x-0.23)^2)+1.5*exp(-2000*(x-0.33)^2)+3*exp(-8000*(x-0.47)^2)+2.25*exp(-16000*(x-0.69)^2)+0.5*exp(-32000*(x-0.83)^2))
mu.s=spike.f(t)
#Gaussian case
mu.t=(1+mu.s)/5
plot(mu.t,type="l")
var.fn=(0.0001+4*(exp(-550*(t-0.2)^2)+exp(-200*(t-0.5)^2)+exp(-950*(t-0.8)^2)))/1.35
plot(var.fn,type="l")
rsnr=sqrt(5)
sigma.t=sqrt(var.fn)/mean(sqrt(var.fn))*sd(mu.t)/rsnr^2
X.s=rnorm(n,mu.t,sigma.t)
mu.est<-ashsmooth.gaus(X.s)
plot(mu.t,type="l")
lines(mu.est,col=2)
#Poisson case
mu.t=0.01+mu.s
X.s=rpois(n,mu.t)
mu.est=ashsmooth.pois(X.s)
plot(mu.t,type="l")
lines(mu.est,col=2)
}
\author{
Matthew Stephens and Zhengrong Xing
}

