% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/ashsmooth.R
\name{ti.thresh}
\alias{ti.thresh}
\title{This function performs TI thresholding for the case where the errors are heteroskedastic.}
\usage{
ti.thresh(x, sigma = NULL, method = "smash", filter.number = 1,
  family = "DaubExPhase", min.level = 3)
}
\arguments{
\item{filter.number,}{family: the wavelet basis to be used.}

\item{x:}{the data. Should be a vector of length a power of 2}

\item{sigma:}{the standard deviation function. Can be provided if known or estimated beforehand.}

\item{method:}{the method to estimate the variance function. Can be "rmad" for running MAD as in Gao (1997), or "smash".}

\item{min.level:}{the primary resolution level.}
}
\value{
a vector of mean estimates
}
\description{
This function performs TI thresholding for the case where the errors are heteroskedastic.
}

